+++ src/main/java/com/weatherwhere/weatherservice/service/date/DateService.java
@@ -1,0 +1,5 @@
+package com.weatherwhere.weatherservice.service.date;
+
+public interface DateService {
+    String[] getDaysAfterToday(int start, int end);
+}
